Organic matter and remineralization

Particulate organic carbon fluxes

PALEOaqchem.RCmultiG.ReactionRCmultiG — Type

ReactionRCmultiG

Discrete representation of organic matter reaction-continuum model from Dale (2015)

The fraction of POC having reactivity k (yr-1) in interval dk at time zero is given by the PDF:

$f(k, 0) dk = \frac{a^{\nu} k^{\nu-1} e^{-a k}}{\Gamma(\nu)} dk$

This is represented by a discrete distribution in n-2 bins spaced logarithmically in k, with additional bins at the start and end to represent POC with reactivity below and above the first and last bin.

This Reaction:

Defines n state variables Corg_<n> with corresponding concentrations Corg_<n>_conc and source-sink Corg_<n>_sms, and also calculates total concentration (summed over POC fractions) Corg_conc.

Applies POC degradation to each discrete fraction and accumulates to variable POC_decay.

If do_rc_distribution = true, distributes input POC defined by Corg_sms into the Corg_<n> state variables, according to the discretised RC continuum distribution, optionally modified by multipliers specified by parameter k_dist_modifier.

References

Dale (2015) GBC https://dx.doi.org/10.1002/2014GB005017

Parameters

k_bin_edges[Vector{Float64}]=[1.0e-10, 1.0e-9, 1.0e-8, 1.0e-7, 1.0e-6, 1.0e-5, 0.0001, 0.001, 0.01, 0.1, 1.0, 10.0, 100.0] (yr-1), default_value=[1.0e-10, 1.0e-9, 1.0e-8, 1.0e-7, 1.0e-6, 1.0e-5, 0.0001, 0.001, 0.01, 0.1, 1.0, 10.0, 100.0], description="n-1 rate-bin edges. Organic matter reactivity is represented an initial bin for k < k_bin_edges[1], n-2 bins centered in log(k), and a final bin for k > bin_edges[end]"

do_RC_distribution[Bool]=true, default_value=true, description="true to transfer Corg_sms to Corg_<n>_sms according to RC initial reactivity distribution, false to just define fractions and decay"

k_dist_modifier[Vector{Float64}]=[1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], default_value=[1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], description="n multipliers for distribution of Corg input to reactivity bins (1.0 to use unmodified RC distribution)"

a[Float64]=0.0003 (yr-1), default_value=0.0003, description="reactive-continuum model average lifetime"

v[Float64]=0.125, default_value=0.125, description="reactive-continuum model shape parameter"

oxidant_dependent_rate[Bool]=false, default_value=false, description="true to scale decomposition rates by factor 'freminOrgTot'"

field_data[DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable isotopes and specify isotope type"

Methods and Variables for default Parameters

do_RCmultiG_conc

volume (m3), VT_ReactDependency, description="cell volume (or cell phase volume eg for a sediment with solid and liquid phases)"

Corg_conc (mol m-3), VT_ReactProperty, description="total Corg concentration from multi G reservoirs"

Corg_1 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_2 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_3 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_4 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_5 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_6 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_7 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_8 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_9 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_10 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_11 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_12 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_13 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_14 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_1_conc (mol m-3), VT_ReactProperty, description="concentration"

Corg_2_conc (mol m-3), VT_ReactProperty, description="concentration"

Corg_3_conc (mol m-3), VT_ReactProperty, description="concentration"

Corg_4_conc (mol m-3), VT_ReactProperty, description="concentration"

Corg_5_conc (mol m-3), VT_ReactProperty, description="concentration"

Corg_6_conc (mol m-3), VT_ReactProperty, description="concentration"

Corg_7_conc (mol m-3), VT_ReactProperty, description="concentration"

Corg_8_conc (mol m-3), VT_ReactProperty, description="concentration"

Corg_9_conc (mol m-3), VT_ReactProperty, description="concentration"

Corg_10_conc (mol m-3), VT_ReactProperty, description="concentration"

Corg_11_conc (mol m-3), VT_ReactProperty, description="concentration"

Corg_12_conc (mol m-3), VT_ReactProperty, description="concentration"

Corg_13_conc (mol m-3), VT_ReactProperty, description="concentration"

Corg_14_conc (mol m-3), VT_ReactProperty, description="concentration"

do_RCmultiG_distribute_input_flux

Corg_sms (mol yr-1), VT_ReactTarget, description="total organic carbon reservoir source - sink (to distribute to RC fractions)"

Corg_1_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

Corg_2_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

Corg_3_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

Corg_4_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

Corg_5_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

Corg_6_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

Corg_7_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

Corg_8_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

Corg_9_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

Corg_10_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

Corg_11_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

Corg_12_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

Corg_13_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

Corg_14_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

do_RCmultiG_POC_decay

POC_decay (mol C yr-1), VT_ReactContributor, description="decay flux"

Corg_1 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_2 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_3 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_4 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_5 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_6 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_7 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_8 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_9 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_10 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_11 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_12 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_13 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_14 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_1_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

Corg_2_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

Corg_3_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

Corg_4_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

Corg_5_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

Corg_6_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

Corg_7_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

Corg_8_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

Corg_9_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

Corg_10_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

Corg_11_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

Corg_12_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

Corg_13_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

Corg_14_sms (mol yr-1), VT_ReactContributor, VF_Deriv, description="organic carbon reservoir discrete RC fraction source - sink"

totals

Corg_4 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_11 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_9 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_1 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_14 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_10 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_2 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_8 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_5 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_12 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_3 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_6 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_7 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_13 (mol), VT_ReactDependency, VF_StateExplicit, description="organic carbon reservoir discrete RC fraction"

Corg_4_total (mol), VT_ReactProperty, description="total organic carbon reservoir discrete RC fraction"

Corg_11_total (mol), VT_ReactProperty, description="total organic carbon reservoir discrete RC fraction"

Corg_9_total (mol), VT_ReactProperty, description="total organic carbon reservoir discrete RC fraction"

Corg_1_total (mol), VT_ReactProperty, description="total organic carbon reservoir discrete RC fraction"

Corg_14_total (mol), VT_ReactProperty, description="total organic carbon reservoir discrete RC fraction"

Corg_10_total (mol), VT_ReactProperty, description="total organic carbon reservoir discrete RC fraction"

Corg_2_total (mol), VT_ReactProperty, description="total organic carbon reservoir discrete RC fraction"

Corg_8_total (mol), VT_ReactProperty, description="total organic carbon reservoir discrete RC fraction"

Corg_5_total (mol), VT_ReactProperty, description="total organic carbon reservoir discrete RC fraction"

Corg_12_total (mol), VT_ReactProperty, description="total organic carbon reservoir discrete RC fraction"

Corg_3_total (mol), VT_ReactProperty, description="total organic carbon reservoir discrete RC fraction"

Corg_6_total (mol), VT_ReactProperty, description="total organic carbon reservoir discrete RC fraction"

Corg_7_total (mol), VT_ReactProperty, description="total organic carbon reservoir discrete RC fraction"

Corg_13_total (mol), VT_ReactProperty, description="total organic carbon reservoir discrete RC fraction"

Remin

PALEOaqchem.Remin.ReactionReminPonly — Type

ReactionReminPonly

Organic particulate matter remineralization (no oxidant use)

Parameters

CIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable carbon isotopes and specify isotope type"

Methods and Variables for default Parameters

do_remin_Ponly

remin_P (mol yr-1), VT_ReactTarget, description="flux P"

remin_N (mol yr-1), VT_ReactTarget, description="flux N"

remin_Corg (mol yr-1), VT_ReactTarget, description="flux Corg"

remin_Ccarb (mol yr-1), VT_ReactTarget, description="flux Ccarb"

soluteflux_P (mol yr-1), VT_ReactContributor, description="flux P"

[soluteflux_TNH3] (mol yr-1), VT_ReactContributor, description="flux TNH3"

[soluteflux_DIC] (mol yr-1), VT_ReactContributor, description="flux DIC"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="flux TAlk"

PALEOaqchem.Remin.ReactionReminO2 — Type

ReactionReminO2

Organic particulate matter remineralization (O2 oxidant only)

Parameters

CIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable carbon isotopes and specify isotope type"

Methods and Variables for default Parameters

do_remin_O2

remin_P (mol yr-1), VT_ReactTarget, description="flux P"

remin_N (mol yr-1), VT_ReactTarget, description="flux N"

remin_Corg (mol yr-1), VT_ReactTarget, description="flux Corg"

remin_Ccarb (mol yr-1), VT_ReactTarget, description="flux Ccarb"

soluteflux_P (mol yr-1), VT_ReactContributor, description="flux P"

[soluteflux_TNH3] (mol yr-1), VT_ReactContributor, description="flux TNH3"

[soluteflux_DIC] (mol yr-1), VT_ReactContributor, description="flux DIC"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="flux TAlk"

reminOrgOxO2 (mol O2 yr-1), VT_ReactProperty, description="oxygen consumption (-ve) by remineralization"

RateStoich_reminOrgOxO2

reminOrgOxO2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

[soluteflux_O2] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxO2"

totals

reminOrgOxO2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

reminOrgOxO2_total (mol O2 yr-1), VT_ReactProperty, description="total oxygen consumption (-ve) by remineralization"

PALEOaqchem.Remin.ReactionReminO2_SO4 — Type

ReactionReminO2_SO4

Organic particulate matter remineralization (O2, SO4 oxidants)

Parameters

oxreminlimit[Float64]=0.008 (mol m-3), default_value=0.008, description="oxygen concentration below which use of O2 for remineralisation is inhibited"

CIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable carbon isotopes and specify isotope type"

SIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable sulphur isotopes and specify isotope type"

Methods and Variables for default Parameters

do_remin_O2_SO4

remin_P (mol yr-1), VT_ReactTarget, description="flux P"

remin_N (mol yr-1), VT_ReactTarget, description="flux N"

remin_Corg (mol yr-1), VT_ReactTarget, description="flux Corg"

remin_Ccarb (mol yr-1), VT_ReactTarget, description="flux Ccarb"

soluteflux_P (mol yr-1), VT_ReactContributor, description="flux P"

[soluteflux_TNH3] (mol yr-1), VT_ReactContributor, description="flux TNH3"

[soluteflux_DIC] (mol yr-1), VT_ReactContributor, description="flux DIC"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="flux TAlk"

O2_conc (mol m-3), VT_ReactDependency, description="O2 concentration"

reminOrgOxO2 (mol O2 yr-1), VT_ReactProperty, description="oxygen consumption (-ve) by remineralization"

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactProperty, description="2 * sulphate consumption (-ve) by remineralization"

RateStoich_reminOrgOxO2

reminOrgOxO2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

[soluteflux_O2] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxO2"

RateStoich_reminOrgOxSO4

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactDependency, description="2 * sulphate consumption (-ve) by remineralization"

[soluteflux_SO4] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

[soluteflux_H2S] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

totals

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactDependency, description="2 * sulphate consumption (-ve) by remineralization"

reminOrgOxO2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

reminOrgOxSO4_total (mol O2eq yr-1), VT_ReactProperty, description="total 2 * sulphate consumption (-ve) by remineralization"

reminOrgOxO2_total (mol O2 yr-1), VT_ReactProperty, description="total oxygen consumption (-ve) by remineralization"

PALEOaqchem.Remin.ReactionReminO2_SO4_CH4 — Type

ReactionReminO2_SO4_CH4

Organic particulate matter remineralization (O2, SO4 oxidants, remaining Corg to CH4)

Parameters

oxreminlimit[Float64]=0.008 (mol m-3), default_value=0.008, description="oxygen concentration below which use of O2 for remineralisation is inhibited"

SO4reminlimit[Float64]=1.0 (mol m-3), default_value=1.0, description="sulphate concentration below which use of SO4 for remineralisation is inhibited"

CIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable carbon isotopes and specify isotope type"

SIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable sulphur isotopes and specify isotope type"

Methods and Variables for default Parameters

do_remin_O2_SO4_CH4

remin_P (mol yr-1), VT_ReactTarget, description="flux P"

remin_N (mol yr-1), VT_ReactTarget, description="flux N"

remin_Corg (mol yr-1), VT_ReactTarget, description="flux Corg"

remin_Ccarb (mol yr-1), VT_ReactTarget, description="flux Ccarb"

soluteflux_P (mol yr-1), VT_ReactContributor, description="flux P"

[soluteflux_TNH3] (mol yr-1), VT_ReactContributor, description="flux TNH3"

[soluteflux_DIC] (mol yr-1), VT_ReactContributor, description="flux DIC"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="flux TAlk"

O2_conc (mol m-3), VT_ReactDependency, description="O2 concentration"

SO4_conc (mol m-3), VT_ReactDependency, description="SO4 concentration"

reminOrgOxO2 (mol O2 yr-1), VT_ReactProperty, description="oxygen consumption (-ve) by remineralization"

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactProperty, description="2 * sulphate consumption (-ve) by remineralization"

reminOrgOxCH4 (mol O2eq yr-1), VT_ReactProperty, description="2 * DIC -> methane (-ve) by remineralization"

RateStoich_reminOrgOxO2

reminOrgOxO2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

[soluteflux_O2] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxO2"

RateStoich_reminOrgOxSO4

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactDependency, description="2 * sulphate consumption (-ve) by remineralization"

[soluteflux_SO4] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

[soluteflux_H2S] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

RateStoich_reminOrgOxCH4

reminOrgOxCH4 (mol O2eq yr-1), VT_ReactDependency, description="2 * DIC -> methane (-ve) by remineralization"

[soluteflux_DIC] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxCH4"

[soluteflux_CH4] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxCH4"

totals

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactDependency, description="2 * sulphate consumption (-ve) by remineralization"

reminOrgOxO2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

reminOrgOxCH4 (mol O2eq yr-1), VT_ReactDependency, description="2 * DIC -> methane (-ve) by remineralization"

reminOrgOxSO4_total (mol O2eq yr-1), VT_ReactProperty, description="total 2 * sulphate consumption (-ve) by remineralization"

reminOrgOxO2_total (mol O2 yr-1), VT_ReactProperty, description="total oxygen consumption (-ve) by remineralization"

reminOrgOxCH4_total (mol O2eq yr-1), VT_ReactProperty, description="total 2 * DIC -> methane (-ve) by remineralization"

PALEOaqchem.Remin.ReactionReminO2_Fe_SO4_CH4 — Type

ReactionReminO2_Fe_SO4_CH4

Organic particulate matter remineralization (O2, FeIIIOx, SO4 oxidants, remaining Corg to CH4)

FeIIIOxreminlimit is from (Van Cappellen and Wang, 1996), 100 umol Fe(OH)3 g-1, assuming dry density is 2.5 g/cm^3 (= 2.5e6 g m^-3) = 100e-6*2.5e6 = 250 mol m-3 solid phase

Parameters

oxreminlimit[Float64]=0.008 (mol m-3), default_value=0.008, description="oxygen concentration below which use of O2 for remineralisation is inhibited"

FeIIIOxreminlimit[Float64]=250.0 (mol m-3), default_value=250.0, description="FeIII oxide concentration below which use of FeIII for remineralisation is inhibited"

SO4reminlimit[Float64]=1.0 (mol m-3), default_value=1.0, description="sulphate concentration below which use of SO4 for remineralisation is inhibited"

CIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable carbon isotopes and specify isotope type"

SIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable sulphur isotopes and specify isotope type"

Methods and Variables for default Parameters

do_remin_O2_Fe_SO4_CH4

remin_P (mol yr-1), VT_ReactTarget, description="flux P"

remin_N (mol yr-1), VT_ReactTarget, description="flux N"

remin_Corg (mol yr-1), VT_ReactTarget, description="flux Corg"

remin_Ccarb (mol yr-1), VT_ReactTarget, description="flux Ccarb"

soluteflux_P (mol yr-1), VT_ReactContributor, description="flux P"

[soluteflux_TNH3] (mol yr-1), VT_ReactContributor, description="flux TNH3"

[soluteflux_DIC] (mol yr-1), VT_ReactContributor, description="flux DIC"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="flux TAlk"

O2_conc (mol m-3), VT_ReactDependency, description="O2 concentration"

FeIIIOx_conc (mol m-3), VT_ReactDependency, description="FeIII oxide concentration"

SO4_conc (mol m-3), VT_ReactDependency, description="SO4 concentration"

reminOrgOxO2 (mol O2 yr-1), VT_ReactProperty, description="oxygen consumption (-ve) by remineralization"

reminOrgOxFeIIIOx (mol O2eq yr-1), VT_ReactProperty, description="oxygen consumption (-ve) by remineralization"

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactProperty, description="2 * sulphate consumption (-ve) by remineralization"

reminOrgOxCH4 (mol O2eq yr-1), VT_ReactProperty, description="2 * DIC -> methane (-ve) by remineralization"

RateStoich_reminOrgOxO2

reminOrgOxO2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

[soluteflux_O2] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxO2"

RateStoich_reminOrgOxFeIIIOx

reminOrgOxFeIIIOx (mol O2eq yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

[soluteflux_FeIIIOx] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxFeIIIOx"

[soluteflux_FeII] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxFeIIIOx"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxFeIIIOx"

RateStoich_reminOrgOxSO4

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactDependency, description="2 * sulphate consumption (-ve) by remineralization"

[soluteflux_SO4] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

[soluteflux_H2S] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

RateStoich_reminOrgOxCH4

reminOrgOxCH4 (mol O2eq yr-1), VT_ReactDependency, description="2 * DIC -> methane (-ve) by remineralization"

[soluteflux_DIC] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxCH4"

[soluteflux_CH4] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxCH4"

totals

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactDependency, description="2 * sulphate consumption (-ve) by remineralization"

reminOrgOxO2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

reminOrgOxCH4 (mol O2eq yr-1), VT_ReactDependency, description="2 * DIC -> methane (-ve) by remineralization"

reminOrgOxFeIIIOx (mol O2eq yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

reminOrgOxSO4_total (mol O2eq yr-1), VT_ReactProperty, description="total 2 * sulphate consumption (-ve) by remineralization"

reminOrgOxO2_total (mol O2 yr-1), VT_ReactProperty, description="total oxygen consumption (-ve) by remineralization"

reminOrgOxCH4_total (mol O2eq yr-1), VT_ReactProperty, description="total 2 * DIC -> methane (-ve) by remineralization"

reminOrgOxFeIIIOx_total (mol O2eq yr-1), VT_ReactProperty, description="total oxygen consumption (-ve) by remineralization"

PALEOaqchem.Remin.ReactionReminO2_NN_Mn_Fe_SO4_CH4 — Type

ReactionReminO2_NN_Mn_Fe_SO4_CH4

Organic particulate matter remineralization (succesive use of O2, NO2, NO3, MnIVOx, FeIIIOx, SO4 oxidants, remaining Corg to CH4)

FeIIIOxreminlimit is from (Van Cappellen and Wang, 1996), 100 umol Fe(OH)3 g-1, assuming dry density is 2.5 g/cm^3 (= 2.5e6 g m^-3) = 100e-6*2.5e6 = 250 mol m-3 solid phase

MnIVOxreminlimit is from (Van Cappellen and Wang, 1996), 16 umol MnO2 g-1, assuming dry density is 2.5 g/cm^3 (= 2.5e6 g m^-3) = 16e-6*2.5e6 = 40 mol m-3 solid phase

Parameters

oxreminlimit[Float64]=0.001 (mol m-3), default_value=0.001, description="oxygen concentration below which use of O2 for remineralisation is inhibited"

NO2reminlimit[Float64]=0.01 (mol m-3), default_value=0.01, description="NO2 concentration below which use of NO2 for remineralisation is inhibited"

NO3reminlimit[Float64]=0.01 (mol m-3), default_value=0.01, description="NO3 concentration below which use of NO3 for remineralisation is inhibited"

MnIVOxreminlimit[Float64]=40.0 (mol m-3), default_value=40.0, description="MnIV oxide concentration below which use of MnIV oxide for remineralisation is inhibited"

FeIIIOxreminlimit[Float64]=250.0 (mol m-3), default_value=250.0, description="FeIII oxide concentration below which use of FeIII oxide for remineralisation is inhibited"

SO4reminlimit[Float64]=0.5 (mol m-3), default_value=0.5, description="sulphate concentration below which use of SO4 for remineralisation is inhibited"

aSO4[Float64]=1.0, default_value=1.0, description="attenuation factor for sulphate reduction"

oxidant_dependent_rate[Bool]=false, default_value=false, description="true to scale decomposition rates by factor 'freminOrgTot'"

CIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable carbon isotopes and specify isotope type"

SIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable sulphur isotopes and specify isotope type"

Methods and Variables for default Parameters

do_remin_O2_NN_Mn_Fe_SO4_CH4_ox_frac

O2_conc (mol m-3), VT_ReactDependency, description="O2 concentration"

NO2_conc (mol m-3), VT_ReactDependency, description="NO2 concentration"

NO3_conc (mol m-3), VT_ReactDependency, description="NO3 concentration"

MnIVOx_conc (mol m-3), VT_ReactDependency, description="MnIV oxide concentration"

FeIIIOx_conc (mol m-3), VT_ReactDependency, description="FeIII oxide concentration"

SO4_conc (mol m-3), VT_ReactDependency, description="SO4 concentration"

freminOrgO2 (), VT_ReactProperty, description="fraction of remineralization by O2"

freminOrgNO2 (), VT_ReactProperty, description="fraction of remineralization by NO2"

freminOrgNO3 (), VT_ReactProperty, description="fraction of remineralization by NO3"

freminOrgMnIVOx (), VT_ReactProperty, description="fraction of remineralization by MnIVOx"

freminOrgFeIIIOx (), VT_ReactProperty, description="fraction of remineralization by FeIIIOx"

freminOrgSO4 (), VT_ReactProperty, description="fraction of remineralization by SO4"

freminOrgCH4 (), VT_ReactProperty, description="fraction of remineralization by CH4"

do_remin_O2_NN_Mn_Fe_SO4_CH4_rates

remin_P (mol yr-1), VT_ReactTarget, description="flux P"

remin_N (mol yr-1), VT_ReactTarget, description="flux N"

remin_Corg (mol yr-1), VT_ReactTarget, description="flux Corg"

remin_Ccarb (mol yr-1), VT_ReactTarget, description="flux Ccarb"

soluteflux_P (mol yr-1), VT_ReactContributor, description="flux P"

[soluteflux_TNH3] (mol yr-1), VT_ReactContributor, description="flux TNH3"

[soluteflux_DIC] (mol yr-1), VT_ReactContributor, description="flux DIC"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="flux TAlk"

freminOrgO2 (), VT_ReactDependency, description="fraction of remineralization by O2"

freminOrgNO2 (), VT_ReactDependency, description="fraction of remineralization by NO2"

freminOrgNO3 (), VT_ReactDependency, description="fraction of remineralization by NO3"

freminOrgMnIVOx (), VT_ReactDependency, description="fraction of remineralization by MnIVOx"

freminOrgFeIIIOx (), VT_ReactDependency, description="fraction of remineralization by FeIIIOx"

freminOrgSO4 (), VT_ReactDependency, description="fraction of remineralization by SO4"

freminOrgCH4 (), VT_ReactDependency, description="fraction of remineralization by CH4"

reminOrgOxO2 (mol O2 yr-1), VT_ReactProperty, description="oxygen consumption (-ve) by remineralization"

reminOrgOxNO2 (mol O2eq yr-1), VT_ReactProperty, description="3/4 NO2 consumption (-ve) by remineralization"

reminOrgOxNO3NO2 (mol O2eq yr-1), VT_ReactProperty, description="1/2 NO3 consumption (-ve) by remineralization"

reminOrgOxMnIVOx (mol O2eq yr-1), VT_ReactProperty, description="oxygen consumption (-ve) by remineralization"

reminOrgOxFeIIIOx (mol O2eq yr-1), VT_ReactProperty, description="oxygen consumption (-ve) by remineralization"

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactProperty, description="2 * sulphate consumption (-ve) by remineralization"

reminOrgOxCH4 (mol O2eq yr-1), VT_ReactProperty, description="2 * DIC -> methane (-ve) by remineralization"

RateStoich_reminOrgOxO2

reminOrgOxO2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

[soluteflux_O2] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxO2"

RateStoich_reminOrgOxNO3NO2

reminOrgOxNO3NO2 (mol O2eq yr-1), VT_ReactDependency, description="1/2 NO3 consumption (-ve) by remineralization"

[soluteflux_NO3] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxNO3NO2"

[soluteflux_NO2] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxNO3NO2"

RateStoich_reminOrgOxNO2

reminOrgOxNO2 (mol O2eq yr-1), VT_ReactDependency, description="3/4 NO2 consumption (-ve) by remineralization"

[soluteflux_NO2] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxNO2"

[soluteflux_N2] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxNO2"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxNO2"

RateStoich_reminOrgOxMnIVOx

reminOrgOxMnIVOx (mol O2eq yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

[soluteflux_MnIVOx] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxMnIVOx"

[soluteflux_MnII] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxMnIVOx"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxMnIVOx"

RateStoich_reminOrgOxFeIIIOx

reminOrgOxFeIIIOx (mol O2eq yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

[soluteflux_FeIIIOx] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxFeIIIOx"

[soluteflux_FeII] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxFeIIIOx"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxFeIIIOx"

RateStoich_reminOrgOxSO4

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactDependency, description="2 * sulphate consumption (-ve) by remineralization"

[soluteflux_SO4] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

[soluteflux_H2S] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

RateStoich_reminOrgOxCH4

reminOrgOxCH4 (mol O2eq yr-1), VT_ReactDependency, description="2 * DIC -> methane (-ve) by remineralization"

[soluteflux_DIC] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxCH4"

[soluteflux_CH4] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxCH4"

totals

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactDependency, description="2 * sulphate consumption (-ve) by remineralization"

reminOrgOxO2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

reminOrgOxNO2 (mol O2eq yr-1), VT_ReactDependency, description="3/4 NO2 consumption (-ve) by remineralization"

reminOrgOxNO3NO2 (mol O2eq yr-1), VT_ReactDependency, description="1/2 NO3 consumption (-ve) by remineralization"

reminOrgOxMnIVOx (mol O2eq yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

reminOrgOxCH4 (mol O2eq yr-1), VT_ReactDependency, description="2 * DIC -> methane (-ve) by remineralization"

reminOrgOxFeIIIOx (mol O2eq yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

reminOrgOxSO4_total (mol O2eq yr-1), VT_ReactProperty, description="total 2 * sulphate consumption (-ve) by remineralization"

reminOrgOxO2_total (mol O2 yr-1), VT_ReactProperty, description="total oxygen consumption (-ve) by remineralization"

reminOrgOxNO2_total (mol O2eq yr-1), VT_ReactProperty, description="total 3/4 NO2 consumption (-ve) by remineralization"

reminOrgOxNO3NO2_total (mol O2eq yr-1), VT_ReactProperty, description="total 1/2 NO3 consumption (-ve) by remineralization"

reminOrgOxMnIVOx_total (mol O2eq yr-1), VT_ReactProperty, description="total oxygen consumption (-ve) by remineralization"

reminOrgOxCH4_total (mol O2eq yr-1), VT_ReactProperty, description="total 2 * DIC -> methane (-ve) by remineralization"

reminOrgOxFeIIIOx_total (mol O2eq yr-1), VT_ReactProperty, description="total oxygen consumption (-ve) by remineralization"

PALEOaqchem.Remin.ReactionReminO2_NO3_Mn_Fe_SO4_CH4 — Type

ReactionReminO2_NO3_Mn_Fe_SO4_CH4

Organic particulate matter remineralization (succesive use of O2, NO3, MnIVOx, FeIIIOx, SO4 oxidants, remaining Corg to CH4)

FeIIIOxreminlimit is from (Van Cappellen and Wang, 1996), 100 umol Fe(OH)3 g-1, assuming dry density is 2.5 g/cm^3 (= 2.5e6 g m^-3) = 100e-6*2.5e6 = 250 mol m-3 solid phase

MnIVOxreminlimit is from (Van Cappellen and Wang, 1996), 16 umol MnO2 g-1, assuming dry density is 2.5 g/cm^3 (= 2.5e6 g m^-3) = 16e-6*2.5e6 = 40 mol m-3 solid phase

Parameters

oxreminlimit[Float64]=0.001 (mol m-3), default_value=0.001, description="oxygen concentration below which use of O2 for remineralisation is inhibited"

NO3reminlimit[Float64]=0.01 (mol m-3), default_value=0.01, description="NO3 concentration below which use of NO3 for remineralisation is inhibited"

MnIVOxreminlimit[Float64]=40.0 (mol m-3), default_value=40.0, description="MnIV oxide concentration below which use of MnIV oxide for remineralisation is inhibited"

FeIIIOxreminlimit[Float64]=250.0 (mol m-3), default_value=250.0, description="FeIII oxide concentration below which use of FeIII oxide for remineralisation is inhibited"

SO4reminlimit[Float64]=0.5 (mol m-3), default_value=0.5, description="sulphate concentration below which use of SO4 for remineralisation is inhibited"

aSO4[Float64]=1.0, default_value=1.0, description="attenuation factor for sulphate reduction"

oxidant_dependent_rate[Bool]=false, default_value=false, description="true to scale decomposition rates by factor 'freminOrgTot'"

CIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable carbon isotopes and specify isotope type"

SIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable sulphur isotopes and specify isotope type"

Methods and Variables for default Parameters

do_remin_O2_NO3_Mn_Fe_SO4_CH4_ox_frac

O2_conc (mol m-3), VT_ReactDependency, description="O2 concentration"

NO3_conc (mol m-3), VT_ReactDependency, description="NO3 concentration"

MnIVOx_conc (mol m-3), VT_ReactDependency, description="MnIV oxide concentration"

FeIIIOx_conc (mol m-3), VT_ReactDependency, description="FeIII oxide concentration"

SO4_conc (mol m-3), VT_ReactDependency, description="SO4 concentration"

freminOrgO2 (), VT_ReactProperty, description="fraction of remineralization by O2"

freminOrgNO3 (), VT_ReactProperty, description="fraction of remineralization by NO3"

freminOrgMnIVOx (), VT_ReactProperty, description="fraction of remineralization by MnIVOx"

freminOrgFeIIIOx (), VT_ReactProperty, description="fraction of remineralization by FeIIIOx"

freminOrgSO4 (), VT_ReactProperty, description="fraction of remineralization by SO4"

freminOrgCH4 (), VT_ReactProperty, description="fraction of remineralization by CH4"

do_remin_O2_NO3_Mn_Fe_SO4_CH4_rates

remin_P (mol yr-1), VT_ReactTarget, description="flux P"

remin_N (mol yr-1), VT_ReactTarget, description="flux N"

remin_Corg (mol yr-1), VT_ReactTarget, description="flux Corg"

remin_Ccarb (mol yr-1), VT_ReactTarget, description="flux Ccarb"

soluteflux_P (mol yr-1), VT_ReactContributor, description="flux P"

[soluteflux_TNH3] (mol yr-1), VT_ReactContributor, description="flux TNH3"

[soluteflux_DIC] (mol yr-1), VT_ReactContributor, description="flux DIC"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="flux TAlk"

freminOrgO2 (), VT_ReactDependency, description="fraction of remineralization by O2"

freminOrgNO3 (), VT_ReactDependency, description="fraction of remineralization by NO3"

freminOrgMnIVOx (), VT_ReactDependency, description="fraction of remineralization by MnIVOx"

freminOrgFeIIIOx (), VT_ReactDependency, description="fraction of remineralization by FeIIIOx"

freminOrgSO4 (), VT_ReactDependency, description="fraction of remineralization by SO4"

freminOrgCH4 (), VT_ReactDependency, description="fraction of remineralization by CH4"

reminOrgOxO2 (mol O2 yr-1), VT_ReactProperty, description="oxygen consumption (-ve) by remineralization"

reminOrgOxNO3only (mol O2eq yr-1), VT_ReactProperty, description="5/4 NO3 consumption (-ve) by remineralization"

reminOrgOxMnIVOx (mol O2eq yr-1), VT_ReactProperty, description="oxygen consumption (-ve) by remineralization"

reminOrgOxFeIIIOx (mol O2eq yr-1), VT_ReactProperty, description="oxygen consumption (-ve) by remineralization"

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactProperty, description="2 * sulphate consumption (-ve) by remineralization"

reminOrgOxCH4 (mol O2eq yr-1), VT_ReactProperty, description="2 * DIC -> methane (-ve) by remineralization"

RateStoich_reminOrgOxO2

reminOrgOxO2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

[soluteflux_O2] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxO2"

RateStoich_reminOrgOxNO3only

reminOrgOxNO3only (mol O2eq yr-1), VT_ReactDependency, description="5/4 NO3 consumption (-ve) by remineralization"

[soluteflux_NO3] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxNO3only"

[soluteflux_N2] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxNO3only"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxNO3only"

RateStoich_reminOrgOxMnIVOx

reminOrgOxMnIVOx (mol O2eq yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

[soluteflux_MnIVOx] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxMnIVOx"

[soluteflux_MnII] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxMnIVOx"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxMnIVOx"

RateStoich_reminOrgOxFeIIIOx

reminOrgOxFeIIIOx (mol O2eq yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

[soluteflux_FeIIIOx] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxFeIIIOx"

[soluteflux_FeII] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxFeIIIOx"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxFeIIIOx"

RateStoich_reminOrgOxSO4

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactDependency, description="2 * sulphate consumption (-ve) by remineralization"

[soluteflux_SO4] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

[soluteflux_H2S] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

RateStoich_reminOrgOxCH4

reminOrgOxCH4 (mol O2eq yr-1), VT_ReactDependency, description="2 * DIC -> methane (-ve) by remineralization"

[soluteflux_DIC] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxCH4"

[soluteflux_CH4] (mol yr-1), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxCH4"

totals

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactDependency, description="2 * sulphate consumption (-ve) by remineralization"

reminOrgOxO2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

reminOrgOxMnIVOx (mol O2eq yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

reminOrgOxCH4 (mol O2eq yr-1), VT_ReactDependency, description="2 * DIC -> methane (-ve) by remineralization"

reminOrgOxNO3only (mol O2eq yr-1), VT_ReactDependency, description="5/4 NO3 consumption (-ve) by remineralization"

reminOrgOxFeIIIOx (mol O2eq yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

reminOrgOxSO4_total (mol O2eq yr-1), VT_ReactProperty, description="total 2 * sulphate consumption (-ve) by remineralization"

reminOrgOxO2_total (mol O2 yr-1), VT_ReactProperty, description="total oxygen consumption (-ve) by remineralization"

reminOrgOxMnIVOx_total (mol O2eq yr-1), VT_ReactProperty, description="total oxygen consumption (-ve) by remineralization"

reminOrgOxCH4_total (mol O2eq yr-1), VT_ReactProperty, description="total 2 * DIC -> methane (-ve) by remineralization"

reminOrgOxNO3only_total (mol O2eq yr-1), VT_ReactProperty, description="total 5/4 NO3 consumption (-ve) by remineralization"

reminOrgOxFeIIIOx_total (mol O2eq yr-1), VT_ReactProperty, description="total oxygen consumption (-ve) by remineralization"