PALEOaqchem Reactions

Particulate fluxes

PALEOaqchem.Particle.ReactionParticleDecay — Type

ReactionParticleDecay

Decay (eg organic matter to remineralization) at rate 1.0/decay_timescale of eg an organic matter dissolved/particulate phase in Reservoir Particle, to decayflux. Particle may have isotope_type. The Reservoir Particle may have the :vsink attribute set to represent a marine sinking particulate phase.

Parameters

decay_timescale[Float64]=0.5 (yr), default_value=0.5, description="particle decay timescale"

decay_threshold[Float64]=-Inf (mol m-3), default_value=-Inf, description="particle decay concentration below which decay stops"

field_data[DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable isotopes and specify isotope type"

Methods and Variables

do_particle_decay

Particle (mol), VT_ReactDependency, description="Particle amount"

Particle_sms (mol yr-1), VT_ReactContributor, description="Particle source-sink"

decayflux (mol yr-1), VT_ReactContributor, description="Particle decay flux"

PALEOaqchem.Particle.ReactionFluxToComponents — Type

ReactionFluxToComponents

Distribute a single input_flux (eg an organic matter flux) to output_fluxnames components with stoichiometry output_fluxstoich. input_flux may have an isotope type (set by field_data) in which case the isotopic composition will be sent to (usually one) output_fluxname with ::Isotope suffix.

Parameters

outputflux_prefix[String]="", default_value="", description="Prefix for output flux names"

outputflux_names[Vector{String}]=["Corg", "N", "P"], default_value=["Corg", "N", "P"], description="Suffixes for output flux names. Use ::Isotope suffix to identify a flux with 'isotope_type'"

outputflux_stoich[Vector{Float64}]=[106.0, 16.0, 1.0], default_value=[106.0, 16.0, 1.0], description="Stoichiometry for output fluxes relative to input flux"

field_data[DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable isotopes and specify isotope type"

Methods and Variables for default Parameters

do_flux_to_components

inputflux (mol m-3), VT_ReactTarget, description="input flux"

[Corg] (mol yr-1), VT_ReactContributor, description="flux Corg"

[N] (mol yr-1), VT_ReactContributor, description="flux N"

[P] (mol yr-1), VT_ReactContributor, description="flux P"

Remin

PALEOaqchem.Remin.ReactionReminPonly — Type

ReactionReminPonly

Organic particulate matter remineralization (no oxidant use)

Parameters

CIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable carbon isotopes and specify isotope type"

Methods and Variables for default Parameters

do_remin_Ponly

remin_P (mol yr-1), VT_ReactTarget, description="flux P"

remin_N (mol yr-1), VT_ReactTarget, description="flux N"

remin_Corg (mol yr-1), VT_ReactTarget, description="flux Corg"

remin_Ccarb (mol yr-1), VT_ReactTarget, description="flux Ccarb"

soluteflux_P (mol yr-1), VT_ReactContributor, description="flux P"

[soluteflux_TNH3] (mol yr-1), VT_ReactContributor, description="flux TNH3"

[soluteflux_DIC] (mol yr-1), VT_ReactContributor, description="flux DIC"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="flux TAlk"

PALEOaqchem.Remin.ReactionReminO2 — Type

ReactionReminO2

Organic particulate matter remineralization (O2 oxidant only)

Parameters

CIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable carbon isotopes and specify isotope type"

Methods and Variables for default Parameters

do_remin_O2

remin_P (mol yr-1), VT_ReactTarget, description="flux P"

remin_N (mol yr-1), VT_ReactTarget, description="flux N"

remin_Corg (mol yr-1), VT_ReactTarget, description="flux Corg"

remin_Ccarb (mol yr-1), VT_ReactTarget, description="flux Ccarb"

soluteflux_P (mol yr-1), VT_ReactContributor, description="flux P"

[soluteflux_TNH3] (mol yr-1), VT_ReactContributor, description="flux TNH3"

[soluteflux_DIC] (mol yr-1), VT_ReactContributor, description="flux DIC"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="flux TAlk"

reminOrgOxO2 (mol O2 yr-1), VT_ReactProperty, description="oxygen consumption (-ve) by remineralization"

RateStoich_reminOrgOxO2

reminOrgOxO2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

[soluteflux_O2] (), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxO2"

totals

reminOrgOxO2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

reminOrgOxO2_total (mol O2 yr-1), VT_ReactProperty, description="total oxygen consumption (-ve) by remineralization"

PALEOaqchem.Remin.ReactionReminO2_SO4 — Type

ReactionReminO2_SO4

Organic particulate matter remineralization (O2, SO4 oxidants)

Parameters

oxreminlimit[Float64]=0.008 (mol m-3), default_value=0.008, description="oxygen concentration below which use of O2 for remineralisation is inhibited"

CIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable carbon isotopes and specify isotope type"

SIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable sulphur isotopes and specify isotope type"

Methods and Variables for default Parameters

do_remin_O2_SO4

remin_P (mol yr-1), VT_ReactTarget, description="flux P"

remin_N (mol yr-1), VT_ReactTarget, description="flux N"

remin_Corg (mol yr-1), VT_ReactTarget, description="flux Corg"

remin_Ccarb (mol yr-1), VT_ReactTarget, description="flux Ccarb"

soluteflux_P (mol yr-1), VT_ReactContributor, description="flux P"

[soluteflux_TNH3] (mol yr-1), VT_ReactContributor, description="flux TNH3"

[soluteflux_DIC] (mol yr-1), VT_ReactContributor, description="flux DIC"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="flux TAlk"

O2_conc (mol m-3), VT_ReactDependency, description="O2 concentration"

reminOrgOxO2 (mol O2 yr-1), VT_ReactProperty, description="oxygen consumption (-ve) by remineralization"

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactProperty, description="2 * sulphate consumption (-ve) by remineralization"

RateStoich_reminOrgOxO2

reminOrgOxO2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

[soluteflux_O2] (), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxO2"

RateStoich_reminOrgOxSO4

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactDependency, description="2 * sulphate consumption (-ve) by remineralization"

[soluteflux_SO4] (), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

[soluteflux_H2S] (), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

[soluteflux_TAlk] (), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

totals

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactDependency, description="2 * sulphate consumption (-ve) by remineralization"

reminOrgOxO2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

reminOrgOxSO4_total (mol O2eq yr-1), VT_ReactProperty, description="total 2 * sulphate consumption (-ve) by remineralization"

reminOrgOxO2_total (mol O2 yr-1), VT_ReactProperty, description="total oxygen consumption (-ve) by remineralization"

PALEOaqchem.Remin.ReactionReminO2_SO4_CH4 — Type

ReactionReminO2_SO4_CH4

Organic particulate matter remineralization (O2, SO4 oxidants, remaining Corg to CH4)

Parameters

oxreminlimit[Float64]=0.008 (mol m-3), default_value=0.008, description="oxygen concentration below which use of O2 for remineralisation is inhibited"

SO4reminlimit[Float64]=1.0 (mol m-3), default_value=1.0, description="sulphate concentration below which use of SO4 for remineralisation is inhibited"

CIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable carbon isotopes and specify isotope type"

SIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable sulphur isotopes and specify isotope type"

Methods and Variables for default Parameters

do_remin_O2_SO4_CH4

remin_P (mol yr-1), VT_ReactTarget, description="flux P"

remin_N (mol yr-1), VT_ReactTarget, description="flux N"

remin_Corg (mol yr-1), VT_ReactTarget, description="flux Corg"

remin_Ccarb (mol yr-1), VT_ReactTarget, description="flux Ccarb"

soluteflux_P (mol yr-1), VT_ReactContributor, description="flux P"

[soluteflux_TNH3] (mol yr-1), VT_ReactContributor, description="flux TNH3"

[soluteflux_DIC] (mol yr-1), VT_ReactContributor, description="flux DIC"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="flux TAlk"

O2_conc (mol m-3), VT_ReactDependency, description="O2 concentration"

SO4_conc (mol m-3), VT_ReactDependency, description="SO4 concentration"

reminOrgOxO2 (mol O2 yr-1), VT_ReactProperty, description="oxygen consumption (-ve) by remineralization"

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactProperty, description="2 * sulphate consumption (-ve) by remineralization"

reminOrgOxCH4 (mol O2eq yr-1), VT_ReactProperty, description="2 * DIC -> methane (-ve) by remineralization"

RateStoich_reminOrgOxO2

reminOrgOxO2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

[soluteflux_O2] (), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxO2"

RateStoich_reminOrgOxSO4

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactDependency, description="2 * sulphate consumption (-ve) by remineralization"

[soluteflux_SO4] (), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

[soluteflux_H2S] (), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

[soluteflux_TAlk] (), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

RateStoich_reminOrgOxCH4

reminOrgOxCH4 (mol O2eq yr-1), VT_ReactDependency, description="2 * DIC -> methane (-ve) by remineralization"

[soluteflux_DIC] (), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxCH4"

[soluteflux_CH4] (), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxCH4"

totals

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactDependency, description="2 * sulphate consumption (-ve) by remineralization"

reminOrgOxO2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

reminOrgOxCH4 (mol O2eq yr-1), VT_ReactDependency, description="2 * DIC -> methane (-ve) by remineralization"

reminOrgOxSO4_total (mol O2eq yr-1), VT_ReactProperty, description="total 2 * sulphate consumption (-ve) by remineralization"

reminOrgOxO2_total (mol O2 yr-1), VT_ReactProperty, description="total oxygen consumption (-ve) by remineralization"

reminOrgOxCH4_total (mol O2eq yr-1), VT_ReactProperty, description="total 2 * DIC -> methane (-ve) by remineralization"

Secondary redox

PALEOaqchem.SecondaryRedox.ReactionRedoxH2S_O2 — Type

ReactionRedoxH2S_O2

Sulphide oxidation by oxygen.

Rate R_H2S_O2	Units	Ref	Notes

1.6e5	(mol l-1)-1 yr-1	(Van Cappellen and Wang, 1996)

3.65e6	(mol l-1)-1 yr-1	(Ozaki et al., 2011)

54e6	(mol l-1)-1 yr-1	(Romaniello and Derry, 2010)

Parameters

R_H2S_O2[Float64]=3650.0 ((mol m-3)-1 yr-1), default_value=3650.0, description="rate constant"

SIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable sulphur isotopes and specify isotope type"

Methods and Variables for default Parameters

do_redox_H2S_O2

redox_H2S_O2 (mol O2 yr-1), VT_ReactProperty, description="oxygen consumption (+ve) by H2S oxidation"

O2_conc (mol m-3), VT_ReactDependency, description="O2 concentration"

H2S_conc (mol m-3), VT_ReactDependency, description="H2S concentration"

volume (m3), VT_ReactDependency, description="box fluid volume"

RateStoich_redox_H2S_O2

redox_H2S_O2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (+ve) by H2S oxidation"

[O2_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_H2S_O2"

[H2S_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_H2S_O2"

[SO4_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_H2S_O2"

[TAlk_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_H2S_O2"

totals

redox_H2S_O2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (+ve) by H2S oxidation"

redox_H2S_O2_total (mol O2 yr-1), VT_ReactProperty, description="total oxygen consumption (+ve) by H2S oxidation"

PALEOaqchem.SecondaryRedox.ReactionRedoxCH4_O2 — Type

ReactionRedoxCH4_O2

Methane oxidation by oxygen.

Rate R_CH4_O2	Units	Ref	Notes

1e10	(mol l-1)-1 yr-1	(Van Cappellen and Wang, 1996)

10e6	(mol l-1)-1 yr-1	TODO

Parameters

R_CH4_O2[Float64]=10000.0 ((mol m-3)-1 yr-1), default_value=10000.0, description="rate constant"

CIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable carbon isotopes and specify isotope type"

Methods and Variables for default Parameters

do_redox_CH4_O2

redox_CH4_O2 (mol O2 yr-1), VT_ReactProperty, description="oxygen consumption (+ve) by CH4 oxidation"

O2_conc (mol m-3), VT_ReactDependency, description="O2 concentration"

CH4_conc (mol m-3), VT_ReactDependency, description="CH4 concentration"

volume (m3), VT_ReactDependency, description="box fluid volume"

RateStoich_redox_CH4_O2

redox_CH4_O2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (+ve) by CH4 oxidation"

[O2_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_CH4_O2"

[CH4_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_CH4_O2"

[DIC_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_CH4_O2"

totals

redox_CH4_O2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (+ve) by CH4 oxidation"

redox_CH4_O2_total (mol O2 yr-1), VT_ReactProperty, description="total oxygen consumption (+ve) by CH4 oxidation"

PALEOaqchem.SecondaryRedox.ReactionRedoxCH4_SO4 — Type

ReactionRedoxCH4_SO4

Methane oxidation by sulphate (anaerobic methane oxidation).

Rate R_CH4_SO4	Units	Ref	Notes

1e4	(mol l-1)-1 yr-1	(Van Cappellen and Wang, 1996)	k17

Parameters

R_CH4_SO4[Float64]=10.0 ((mol m-3)-1 yr-1), default_value=10.0, description="rate constant"

CIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable carbon isotopes and specify isotope type"

Methods and Variables for default Parameters

do_redox_CH4_SO4

redox_CH4_SO4 (mol CH4/SO4 yr-1), VT_ReactProperty, description="sulphate consumption (+ve) by CH4 oxidation"

SO4_conc (mol m-3), VT_ReactDependency, description="SO4 concentration"

CH4_conc (mol m-3), VT_ReactDependency, description="CH4 concentration"

volume (m3), VT_ReactDependency, description="box fluid volume"

RateStoich_redox_CH4_SO4

redox_CH4_SO4 (mol CH4/SO4 yr-1), VT_ReactDependency, description="sulphate consumption (+ve) by CH4 oxidation"

[SO4_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_CH4_SO4"

[CH4_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_CH4_SO4"

[DIC_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_CH4_SO4"

[H2S_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_CH4_SO4"

[TAlk_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_CH4_SO4"

totals

redox_CH4_SO4 (mol CH4/SO4 yr-1), VT_ReactDependency, description="sulphate consumption (+ve) by CH4 oxidation"

redox_CH4_SO4_total (mol CH4/SO4 yr-1), VT_ReactProperty, description="total sulphate consumption (+ve) by CH4 oxidation"

PALEOaqchem.SecondaryRedox.ReactionRedoxFeII_O2 — Type

ReactionRedoxFeII_O2

FeII oxidation by oxygen, producing generic FeIIIOx (0.5 Fe2O3)

4 Fe++ + O2 + 4 H2O -> 4 FeIIIOx + 8 H+

Parameters

R_FeII_O2[Float64]=3650.0 ((mol m-3)-1 yr-1), default_value=3650.0, description="rate constant"

Methods and Variables for default Parameters

do_redox_FeII_O2

redox_FeII_O2 (mol O2 yr-1), VT_ReactProperty, description="oxygen consumption (+ve) by Fe++ oxidation"

O2_conc (mol m-3), VT_ReactDependency, description="O2 concentration"

FeII_conc (mol m-3), VT_ReactDependency, description="FeII concentration"

volume (m3), VT_ReactDependency, description="box fluid volume"

RateStoich_redox_FeII_O2

redox_FeII_O2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (+ve) by Fe++ oxidation"

[O2_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_FeII_O2"

[FeII_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_FeII_O2"

[FeIIIOx_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_FeII_O2"

[TAlk_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_FeII_O2"

totals

redox_FeII_O2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (+ve) by Fe++ oxidation"

redox_FeII_O2_total (mol O2 yr-1), VT_ReactProperty, description="total oxygen consumption (+ve) by Fe++ oxidation"

PALEOaqchem.SecondaryRedox.ReactionRedoxFeIII_H2S — Type

ReactionRedoxFeIII_H2S

Sulphate-mediated iron reduction (of generic iron oxide FeIIIOx ~ 0.5 Fe2O3)

H2S + 8 FeIIIOx + 14 H+ -> 8 Fe++ + SO4-- + 8 H2O

NB: rate constant k [units (mol m-3)^-0.5 yr-1] = 277.2 * k [units (mol l-1)^-0.5 hr-1]

where van de Velde (2021) give k_SMIs = 1.98e0 (mol l-1)^-0.5 hr-1 for solid-phase FeIII

Parameters

R_FeIII_H2S[Float64]=548.856 ((mol m-3)^-0.5 yr-1), default_value=548.856, description="rate constant"

Methods and Variables for default Parameters

do_FeIII_H2S_rate

redox_FeIII_H2S (mol H2S yr-1), VT_ReactProperty, description="sulphide consumption (+ve) by Fe+++ reduction"

H2S_conc (mol m-3), VT_ReactDependency, description="H2S concentration"

FeIIIOx_conc (mol m-3), VT_ReactDependency, description="FeIIIOx concentration"

volume (m3), VT_ReactDependency, description="box fluid volume"

RateStoich_redox_FeIII_H2S

redox_FeIII_H2S (mol H2S yr-1), VT_ReactDependency, description="sulphide consumption (+ve) by Fe+++ reduction"

[H2S_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_FeIII_H2S"

[FeIIIOx_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_FeIII_H2S"

[FeII_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_FeIII_H2S"

[SO4_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_FeIII_H2S"

[TAlk_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_FeIII_H2S"

totals

redox_FeIII_H2S (mol H2S yr-1), VT_ReactDependency, description="sulphide consumption (+ve) by Fe+++ reduction"

redox_FeIII_H2S_total (mol H2S yr-1), VT_ReactProperty, description="total sulphide consumption (+ve) by Fe+++ reduction"

PALEOaqchem.SecondaryRedox.ReactionRedoxFeS_O2 — Type

ReactionRedoxFeS_O2

FeS oxidation

FeS + 2 O2 -> Fe++ + SO4--

Parameters

R_FeS_O2[Float64]=100.0 ((mol m-3)-1 yr-1), default_value=100.0, description="rate constant"

Methods and Variables for default Parameters

do_redox_FeS_O2

redox_FeS_O2 (mol FeS yr-1), VT_ReactProperty, description="FeS consumption (+ve) by oxidation by O2"

FeS_conc (mol m-3), VT_ReactDependency, description="FeS concentration"

O2_conc (mol m-3), VT_ReactDependency, description="O2 concentration"

volume (m3), VT_ReactDependency, description="box fluid volume"

RateStoich_redox_FeS_O2

redox_FeS_O2 (mol FeS yr-1), VT_ReactDependency, description="FeS consumption (+ve) by oxidation by O2"

[FeS_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_FeS_O2"

[O2_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_FeS_O2"

[FeII_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_FeS_O2"

[SO4_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_FeS_O2"

totals

redox_FeS_O2 (mol FeS yr-1), VT_ReactDependency, description="FeS consumption (+ve) by oxidation by O2"

redox_FeS_O2_total (mol FeS yr-1), VT_ReactProperty, description="total FeS consumption (+ve) by oxidation by O2"

PALEOaqchem.SecondaryRedox.ReactionRedoxFeS2pyr_O2 — Type

ReactionRedoxFeS2pyr_O2

Pyrite oxidation

FeS2 + 3.5 O2 + H2O -> Fe++ + 2 SO4-- + 2 H+

Parameters

R_FeS2pyr_O2[Float64]=1.0 ((mol m-3)-1 yr-1), default_value=1.0, description="rate constant"

Methods and Variables for default Parameters

do_redox_FeS2pyr_O2

redox_FeS2pyr_O2 (mol FeS2pyr yr-1), VT_ReactProperty, description="pyrite consumption (+ve) by oxidation by O2"

FeS2pyr_conc (mol m-3), VT_ReactDependency, description="FeS2pyr (pyrite) concentration"

O2_conc (mol m-3), VT_ReactDependency, description="O2 concentration"

volume (m3), VT_ReactDependency, description="box fluid volume"

RateStoich_redox_FeS2pyr_O2

redox_FeS2pyr_O2 (mol FeS2pyr yr-1), VT_ReactDependency, description="pyrite consumption (+ve) by oxidation by O2"

[FeS2pyr_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_FeS2pyr_O2"

[O2_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_FeS2pyr_O2"

[FeII_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_FeS2pyr_O2"

[SO4_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_FeS2pyr_O2"

[TAlk_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_FeS2pyr_O2"

totals

redox_FeS2pyr_O2 (mol FeS2pyr yr-1), VT_ReactDependency, description="pyrite consumption (+ve) by oxidation by O2"

redox_FeS2pyr_O2_total (mol FeS2pyr yr-1), VT_ReactProperty, description="total pyrite consumption (+ve) by oxidation by O2"

Iron-sulphur system

PALEOaqchem.FeS.ReactionFeSaq — Type

ReactionFeSaq

Equilibrium chemistry of FeII, H2S, FeS system

Represents two equilibrium reactions as algebraic constraints to define primary species [Fe2++] and [HS-] in terms of:

total [S-II] = [HS-] + [H2S] + [FeSaq]

total [FeII] = [Fe++] + [FeSaq]

given a provided fixed value of [H+]

H2S        <--> H+ + HS-         eqb. const. K_H2S
FeSaq + H+ <--> Fe++ + HS-       eqb. const K2_FeSaq

Species concentration	PALEO variable name	PALEO variable type

[H+]	Hp_conc	Dependency

total [S-II]	SmIIaqtot_conc	Dependency

[HS-]	Hsm_conc	State

constraint on total [S-II]	SmIIaqtot_constraint	Constraint

[H2S]	H2Ssp_conc	Property

[HS-] + [H2S]	H2Stot_conc	Property

total [FeII]	FeIIaqtot_conc	Dependency

[Fe2++]	FeII_conc	State

constraint on total [FeII]	FeIIaqtot_constraint	Constraint

(usually SmIIaqtot and FeIIaqtot are ODE variables and should be defined by ReactionReservoir)

See:

Rickard (2006) GCA https://10.1016/j.gca.2006.02.029

Rickard and Luther (2007) Chemical Reviews https://dx.doi.org/10.1021/cr0503658

Lenton and Daines (2017) Ann. Rev. Mar. Sci. https://dx.doi.org/10.1146/annurev-marine-010816-060521

van de Velde etal (2021) GMD https://10.5194/gmd-14-2713-2021

Parameters

K_H2S[Float64]=2.15e-7 (mol kg-1), default_value=2.15e-7, description="K_H2S = [H+][HS-]/[H2S] units mol kg-1 eqb const for H2S <–> HS- + H+ (Hofmann (2010) Aquatic Geochemistry https://10.1007/s10498-009-9084-1)"

K2_FeSaq[Float64]=158.48931924611142, default_value=158.48931924611142, description="K2_FeSaq = {Fe++}[HS-]/([FeSaq][H+]) eqb constant for FeSaq + H+ <–> Fe++ + HS- (Rickard (2006) eq 8)"

activity_Fe2p[Float64]=0.23, default_value=0.23, description="Fe++ activity (eg van de Velde 2012 Table 3 from Davies eqn)"

Methods and Variables for default Parameters

do_FeSaq_eqb

HSm_conc (mol m-3), VT_ReactDependency, VF_State, description="aqueous HS- concentration"

FeII_conc (mol m-3), VT_ReactDependency, VF_State, description="aqueous Fe++ concentration"

SmIIaqtot_constraint (mol m-3), VT_ReactContributor, VF_Constraint, description="algebraic constraint on SmIIaqtot_conc (= 0)"

FeIIaqtot_constraint (mol m-3), VT_ReactContributor, VF_Constraint, description="algebraic constraint on FeIIaqtot_conc (= 0)"

Hp_conc (mol m-3), VT_ReactDependency, description="aqueous H+ concentration"

rho_ref (kg m-3), VT_ReactDependency, description="reference density"

SmIIaqtot_conc (mol m-3), VT_ReactDependency, description="total aqueous S-II concentration"

FeIIaqtot_conc (mol m-3), VT_ReactDependency, description="total aqueous FeII concentration"

H2Ssp_conc (mol m-3), VT_ReactProperty, description="aqueous H2S species concentration"

H2Stot_conc (mol m-3), VT_ReactProperty, description="aqueous H2S + HS- concentration"

FeSaq_conc (mol m-3), VT_ReactProperty, description="aqueous FeS concentration"

PALEOaqchem.FeS.ReactionFeSm — Type

ReactionFeSm

FeS precipitation/dissolution

Represents

[FeSaq] <--> [FeSm]

as a fast precipitation-dissolution reaction, using the formulation from Van Cappellen & Wang (1996)

See:

Van Cappellen & Wang (1996) AJS https://dx.doi.org/10.2475/ajs.296.3.197

Rickard (2006) GCA https://dx.doi.org/10.1016/j.gca.2006.02.029

Rickard and Luther (2007) Chemical Reviews https://dx.doi.org/10.1021/cr0503658

Parameters

K0_FeSm[Float64]=1.9952623149688787e-6 (mol l-1), default_value=1.9952623149688787e-6, description="solubility constant for FeSm"

rate_FeS_prec[Float64]=40.0 (mol m-3 yr-1), default_value=40.0, description="rate of FeS precipitation"

rate_FeS_diss[Float64]=1000.0 (yr-1), default_value=1000.0, description="rate of FeS dissolution"

Methods and Variables for default Parameters

do_FeSm

rate_FeSm (mol yr-1), VT_ReactProperty, description="FeSm precipitation - dissolution rate"

volume_solute (m3), VT_ReactDependency, description="solute fluid volume"

volume_solid (m3), VT_ReactDependency, description="solid-phase volume"

FeSaq_conc (mol m-3), VT_ReactDependency, description="aqueous FeS concentration"

Omega_FeSaq (), VT_ReactProperty, description="aqueous FeS saturation"

FeSm_conc (mol m-3), VT_ReactDependency, description="FeS solid phase concentration"

RateStoich_rate_FeSm

rate_FeSm (mol yr-1), VT_ReactDependency, description="FeSm precipitation - dissolution rate"

[SmIIaqtot_sms] (), VT_ReactContributor, description="generated by RateStoich rate=rate_FeSm"

[FeIIaqtot_sms] (), VT_ReactContributor, description="generated by RateStoich rate=rate_FeSm"

[FeSm_sms] (), VT_ReactContributor, description="generated by RateStoich rate=rate_FeSm"

totals

rate_FeSm (mol yr-1), VT_ReactDependency, description="FeSm precipitation - dissolution rate"

rate_FeSm_total (mol yr-1), VT_ReactProperty, description="total FeSm precipitation - dissolution rate"

PALEOaqchem.FeS.ReactionPyrH2S — Type

ReactionPyrH2S

Pyrite formation by the "H2S" (Berzelius) mechanism

FeSm + H2S -> FeS2pyr + H2

with rate

R_Pyr_H2S * [FeSm] * [H2S]

Where RPyrH2S is highly uncertain:

(Dale et al., 2015) use RPyrH2S = 1e5 M-1 yr-1 (= 1e5 *1e-3 = 100 (mol m-3-1) yr-1)

(Van De Velde et al., 2021) use RPyrH2S = 0.3708 M-1 h-1 (= 0.3708 * 8.766 (mol m-3)-1 yr-1), where [H2S] is taken to be total H2S (HS- + H2S)

Parameters

R_Pyr_H2S[Float64]=3.2504328 ((mol m-3)-1 yr-1), default_value=3.2504328, description="rate constant"

Methods and Variables for default Parameters

do_PyrH2S

rate_FeS2pyr_H2S (mol yr-1), VT_ReactProperty, description="pyrite formation rate via the H2S mechanism"

FeSm_conc (mol m-3), VT_ReactDependency, description="FeSm concentration"

H2Ssp_conc (mol m-3), VT_ReactDependency, description="H2S (species) concentration"

volume (m3), VT_ReactDependency, description="box fluid volume"

RateStoich_rate_FeS2pyr_H2S

rate_FeS2pyr_H2S (mol yr-1), VT_ReactDependency, description="pyrite formation rate via the H2S mechanism"

[FeSm_sms] (), VT_ReactContributor, description="generated by RateStoich rate=rate_FeS2pyr_H2S"

[SmIIaqtot_sms] (), VT_ReactContributor, description="generated by RateStoich rate=rate_FeS2pyr_H2S"

[FeS2pyr_sms] (), VT_ReactContributor, description="generated by RateStoich rate=rate_FeS2pyr_H2S"

[H2_sms] (), VT_ReactContributor, description="generated by RateStoich rate=rate_FeS2pyr_H2S"

totals

rate_FeS2pyr_H2S (mol yr-1), VT_ReactDependency, description="pyrite formation rate via the H2S mechanism"

rate_FeS2pyr_H2S_total (mol yr-1), VT_ReactProperty, description="total pyrite formation rate via the H2S mechanism"

Carbonate chemistry

PALEOaqchem.CarbChem.ReactionCO2SYS — Type

ReactionCO2SYS

Carbonate chemistry using PALEOcarbchem.

Three different pH solver approaches are supported, set by Parameter solve_pH:

solve: iterative solution for pH, TAlk is a StateExplicit Variable (eg using explicit ODE solver)

constraint: additional State Variable for pH and algebraic Constraint for TAlk (requires DAE solver).

implicit: TAlk is a Total Variable, a function of a State Variable for pH (requires DAE solver)

Parameters

WhichKs[Int64]=10, default_value=10, description="PALEOcarbchem WhichKs choice of eqb constant data"

WhoseKSO4[Int64]=1, default_value=1, description="PALEOcarbchem WhoseKSO4 choice of eqb constant data"

components[Vector{String}]=["Ci", "S", "F", "B", "Omega"], default_value=["Ci", "S", "F", "B", "Omega"], allowed_values=["Ci", "S", "F", "P", "B", "Si", "H2S", "NH3", "Omega"], description="PALEOcarbchem optional contributions to TAlk"

defaultconcs[Vector{String}]=["TS", "TF", "TB", "Ca"], default_value=["TS", "TF", "TB", "Ca"], allowed_values=["TS", "TF", "TB", "Ca"], description="PALEOcarbchem default concentrations (modern values calculated from salinity)"

outputs[Vector{String}]=["pCO2", "xCO2dryinp"], default_value=["pCO2", "xCO2dryinp"], description="PALEOcarbchem output concentrations etc to include as output variables"

output_pHtot[Bool]=true, default_value=true, description="true to output pHtot (requires TS_conc and TF_conc)"

solve_pH[String]="solve", default_value="solve", allowed_values=["solve", "constraint", "implicit"], description="'solve' to solve iteratively for pH, 'constraint' to include [pHfree] as a state variable and TAlk as a constraint, 'implicit' to include [pHfree] as a state variable and TAlk an implicit variable"

pHtol[Float64]=2.220446049250313e-14, default_value=2.220446049250313e-14, description="pH tolerance for iterative solution"

simd_width[String]="1", default_value="1", allowed_values=["1", "FP64P2", "FP64P4", "FP32P4", "FP32P8"], description="use SIMD ("1" - no SIMD, "FP64P4" - 4 x Float64, etc)"

Methods and Variables for default Parameters

do_modern_default_concs

volume (m^3), VT_ReactDependency, description="Cell volume"

rho_ref (kg m^-3), VT_ReactDependency, description="density conversion factor"

temp (Kelvin), VT_ReactDependency, description="temperature"

pressure (dbar), VT_ReactDependency, description="pressure"

sal (psu), VT_ReactDependency, description="salinity"

TS_conc (mol m^-3), VT_ReactProperty, description="TS_conc modern default scaled from salinity"

TF_conc (mol m^-3), VT_ReactProperty, description="TF_conc modern default scaled from salinity"

TB_conc (mol m^-3), VT_ReactProperty, description="TB_conc modern default scaled from salinity"

Ca_conc (mol m^-3), VT_ReactProperty, description="Ca_conc modern default scaled from salinity"

do_carbchem

volume (m^3), VT_ReactDependency, description="Cell volume"

rho_ref (kg m^-3), VT_ReactDependency, description="density conversion factor"

temp (Kelvin), VT_ReactDependency, description="temperature"

pressure (dbar), VT_ReactDependency, description="pressure"

sal (psu), VT_ReactDependency, description="salinity"

pHfree (), VT_ReactProperty, description="-log10(hydrogen ion concentration)"

TAlk_conc (mol m^-3), VT_ReactDependency, description="total TAlk concentration"

pHtot (), VT_ReactProperty, description="-log10([H] + [HS]), pH on total scale"

TCi_conc (mol m^-3), VT_ReactDependency, description="TCi_concinput concentration"

TS_conc (mol m^-3), VT_ReactDependency, description="TS_concinput concentration"

TF_conc (mol m^-3), VT_ReactDependency, description="TF_concinput concentration"

TB_conc (mol m^-3), VT_ReactDependency, description="TB_concinput concentration"

Ca_conc (mol m^-3), VT_ReactDependency, description="Ca_concinput concentration"

pCO2 –> %reaction%pCO2 (atm), VT_ReactProperty, description="CO2 partial pressure"

xCO2dryinp –> %reaction%xCO2dryinp (), VT_ReactProperty, description="mixing ratio of CO2 in dry air at 1 atm (always > pCO2 due to H2O vapour pressure)"

Isotope systems

Boron

PALEOaqchem.Boron.ReactionBoronIsotope — Type

ReactionBoronIsotope

Calculate d11B for aqueous B(OH)4- and B(OH)3 species from mass balance, given d11B for total B, total B concentration B_conc and B(OH)4- concentration BOH4_conc.

See eg (Zeebe and Wolf-Gladrow, 2001), p220.

Parameters

alphaB[Float64]=1.0272, default_value=1.0272, description="isotopic fractionation factor B(OH)4m <-> B(OH)3"

BIsotope[external, UnionAll]=PALEOboxes.IsotopeLinear, default_value=PALEOboxes.IsotopeLinear, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable boron isotopes and specify isotope type"

Methods and Variables

do_boron_isotope

B_conc (mol m-3), VT_ReactDependency, description="B concentration"

BOH4_conc (mol m-3), VT_ReactDependency, description="B(OH)4m concentration"

B_delta (per mil), VT_ReactDependency, description="d11B delta for total B"

BOH4_delta (per mil), VT_ReactProperty, description="d11B delta for B(OH)4- species"