PALEOaqchem functions

PALEOaqchem.O2AlkUptakeRemin — Function

O2AlkUptakeRemin(Corg, (NO3, TNH3, Ngas), TPO4, Ccarb; rO2Corg=1) -> (O2, Alk)

Oxygen and alkalinity assimilated for production (or released by remineralisation) of particulate matter with specified Corg, Ccarb from specified nitrate NO3, total ammonia TNH3, gaseous N, total phosphate TPO4

NB sign: for Corg +ve, O2 is -ve (ie quantity to subtract from tracer sms for production, or add for remineralisation)

Examples

julia> PALEOaqchem.O2AlkUptakeRemin(106.0, (0.0, 0.0, 0.0), 1.0, 0.0) # Corg:P = 106:1, no N
(-106.0, 0.0)

julia> PALEOaqchem.O2AlkUptakeRemin(106.0, (16.0, 0.0, 0.0), 1.0, 0.0) # Corg:NO3:P = 106:16:1
(-138.0, -16.0)

julia> PALEOaqchem.O2AlkUptakeRemin(106.0, (0.0, 16.0, 0.0), 1.0, 0.0) # Corg:TNH3:P = 106:16:1
(-106.0, 16.0)

PALEOaqchem.MolecularDiffusion.create_solute_diffusivity_func — Function

create_solute_diffusivity_func(species_or_constant::AbstractString) -> f_diffcoeff

Molecular diffusivity for species_or_constant in water

Based on data compiled by (Boudreau, 1997) Ch. 4

If species_or_constant is a species name, returns a function f_diffcoeff, which calculates:

f_diffcoeff(TdegK, Pbar, S) -> Dmol (cm^2/s)

If species_or_constant can be parsed into a Float64, returns a function which provides this constant value (units cm^2/s).

The effect of pressure and salinity is approximated by assuming viscosity (calculated by MolecularDiffusion.sw_dynamic_viscosity) is the primary consideration, (Boudreau, 1997) eqn. 4.107 and Table 4.10.

PALEOaqchem.MolecularDiffusion.sw_dynamic_viscosity — Function

sw_dynamic_viscosity(TdegC, Pbar, S) -> mu

Seawater dynamic viscosity ((Boudreau, 1997) p94) 'claimed to accurate to within 7% over T (0 - 30C), P (1 - 1000 bar), S (0 36)'

Returns mu (10^-2 g/cm/s) (centipoise)

Arguments:

T: deg C (0 - 30C) temperature

P: bar (1 1000) pressure

S: salinity (0 - 36)

See check values for salinity effect Table 4.10 p.125