PALEOaqchem Reactions

Particulate fluxes

PALEOaqchem.Particle.ReactionParticleDecay — Type

ReactionParticleDecay

Decay (eg organic matter to remineralization) at rate 1.0/decay_timescale of eg an organic matter dissolved/particulate phase in Reservoir Particle, to decayflux. Particle may have isotope_type. The Reservoir Particle may have the :vsink attribute set to represent a marine sinking particulate phase.

Parameters

decay_timescale[Float64]=0.5 (yr), default_value=0.5, description="particle decay timescale"

decay_threshold[Float64]=-Inf (mol m-3), default_value=-Inf, description="particle decay concentration below which decay stops"

field_data[DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable isotopes and specify isotope type"

Methods and Variables

do_particle_decay

Particle (mol), VT_ReactDependency, description="Particle amount"

Particle_sms (mol yr-1), VT_ReactContributor, description="Particle source-sink"

decayflux (mol yr-1), VT_ReactContributor, description="Particle decay flux"

PALEOaqchem.Particle.ReactionFluxToComponents — Type

ReactionFluxToComponents

Distribute a single input_flux (eg an organic matter flux) to output_fluxnames components with stoichiometry output_fluxstoich. input_flux may have an isotope type (set by field_data) in which case the isotopic composition will be sent to (usually one) output_fluxname with ::Isotope suffix.

Parameters

outputflux_prefix[String]="", default_value="", description="Prefix for output flux names"

outputflux_names[Vector{String}]=["Corg", "N", "P"], default_value=["Corg", "N", "P"], description="Suffixes for output flux names. Use ::Isotope suffix to identify a flux with 'isotope_type'"

outputflux_stoich[Vector{Float64}]=[106.0, 16.0, 1.0], default_value=[106.0, 16.0, 1.0], description="Stoichiometry for output fluxes relative to input flux"

field_data[DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable isotopes and specify isotope type"

Methods and Variables for default Parameters

do_flux_to_components

inputflux (mol m-3), VT_ReactTarget, description="input flux"

[Corg] (mol yr-1), VT_ReactContributor, description="flux Corg"

[N] (mol yr-1), VT_ReactContributor, description="flux N"

[P] (mol yr-1), VT_ReactContributor, description="flux P"

Remin

PALEOaqchem.Remin.ReactionReminPonly — Type

ReactionReminPonly

Organic particulate matter remineralization (no oxidant use)

Parameters

CIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable carbon isotopes and specify isotope type"

Methods and Variables for default Parameters

do_remin_Ponly

remin_P (mol yr-1), VT_ReactTarget, description="flux P"

remin_N (mol yr-1), VT_ReactTarget, description="flux N"

remin_Corg (mol yr-1), VT_ReactTarget, description="flux Corg"

remin_Ccarb (mol yr-1), VT_ReactTarget, description="flux Ccarb"

soluteflux_P (mol yr-1), VT_ReactContributor, description="flux P"

[soluteflux_TNH3] (mol yr-1), VT_ReactContributor, description="flux TNH3"

[soluteflux_DIC] (mol yr-1), VT_ReactContributor, description="flux DIC"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="flux TAlk"

PALEOaqchem.Remin.ReactionReminO2 — Type

ReactionReminO2

Organic particulate matter remineralization (O2 oxidant only)

Parameters

CIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable carbon isotopes and specify isotope type"

Methods and Variables for default Parameters

do_remin_O2

remin_P (mol yr-1), VT_ReactTarget, description="flux P"

remin_N (mol yr-1), VT_ReactTarget, description="flux N"

remin_Corg (mol yr-1), VT_ReactTarget, description="flux Corg"

remin_Ccarb (mol yr-1), VT_ReactTarget, description="flux Ccarb"

soluteflux_P (mol yr-1), VT_ReactContributor, description="flux P"

[soluteflux_TNH3] (mol yr-1), VT_ReactContributor, description="flux TNH3"

[soluteflux_DIC] (mol yr-1), VT_ReactContributor, description="flux DIC"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="flux TAlk"

reminOrgOxO2 (mol O2 yr-1), VT_ReactProperty, description="oxygen consumption (-ve) by remineralization"

RateStoich_reminOrgOxO2

reminOrgOxO2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

[soluteflux_O2] (), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxO2"

totals

reminOrgOxO2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

reminOrgOxO2_total (mol O2 yr-1), VT_ReactProperty, description="total oxygen consumption (-ve) by remineralization"

PALEOaqchem.Remin.ReactionReminO2_SO4 — Type

ReactionReminO2_SO4

Organic particulate matter remineralization (O2, SO4 oxidants)

Parameters

oxreminlimit[Float64]=0.008 (mol m-3), default_value=0.008, description="oxygen concentration below which use of O2 for remineralisation is inhibited"

CIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable carbon isotopes and specify isotope type"

SIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable sulphur isotopes and specify isotope type"

Methods and Variables for default Parameters

do_remin_O2_SO4

remin_P (mol yr-1), VT_ReactTarget, description="flux P"

remin_N (mol yr-1), VT_ReactTarget, description="flux N"

remin_Corg (mol yr-1), VT_ReactTarget, description="flux Corg"

remin_Ccarb (mol yr-1), VT_ReactTarget, description="flux Ccarb"

soluteflux_P (mol yr-1), VT_ReactContributor, description="flux P"

[soluteflux_TNH3] (mol yr-1), VT_ReactContributor, description="flux TNH3"

[soluteflux_DIC] (mol yr-1), VT_ReactContributor, description="flux DIC"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="flux TAlk"

O2_conc (mol m-3), VT_ReactDependency, description="O2 concentration"

reminOrgOxO2 (mol O2 yr-1), VT_ReactProperty, description="oxygen consumption (-ve) by remineralization"

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactProperty, description="2 * sulphate consumption (-ve) by remineralization"

RateStoich_reminOrgOxO2

reminOrgOxO2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

[soluteflux_O2] (), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxO2"

RateStoich_reminOrgOxSO4

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactDependency, description="2 * sulphate consumption (-ve) by remineralization"

[soluteflux_SO4] (), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

[soluteflux_H2S] (), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

[soluteflux_TAlk] (), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

totals

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactDependency, description="2 * sulphate consumption (-ve) by remineralization"

reminOrgOxO2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

reminOrgOxSO4_total (mol O2eq yr-1), VT_ReactProperty, description="total 2 * sulphate consumption (-ve) by remineralization"

reminOrgOxO2_total (mol O2 yr-1), VT_ReactProperty, description="total oxygen consumption (-ve) by remineralization"

PALEOaqchem.Remin.ReactionReminO2_SO4_CH4 — Type

ReactionReminO2_SO4_CH4

Organic particulate matter remineralization (O2, SO4 oxidants, remaining Corg to CH4)

Parameters

oxreminlimit[Float64]=0.008 (mol m-3), default_value=0.008, description="oxygen concentration below which use of O2 for remineralisation is inhibited"

SO4reminlimit[Float64]=1.0 (mol m-3), default_value=1.0, description="sulphate concentration below which use of SO4 for remineralisation is inhibited"

CIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable carbon isotopes and specify isotope type"

SIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable sulphur isotopes and specify isotope type"

Methods and Variables for default Parameters

do_remin_O2_SO4_CH4

remin_P (mol yr-1), VT_ReactTarget, description="flux P"

remin_N (mol yr-1), VT_ReactTarget, description="flux N"

remin_Corg (mol yr-1), VT_ReactTarget, description="flux Corg"

remin_Ccarb (mol yr-1), VT_ReactTarget, description="flux Ccarb"

soluteflux_P (mol yr-1), VT_ReactContributor, description="flux P"

[soluteflux_TNH3] (mol yr-1), VT_ReactContributor, description="flux TNH3"

[soluteflux_DIC] (mol yr-1), VT_ReactContributor, description="flux DIC"

[soluteflux_TAlk] (mol yr-1), VT_ReactContributor, description="flux TAlk"

O2_conc (mol m-3), VT_ReactDependency, description="O2 concentration"

SO4_conc (mol m-3), VT_ReactDependency, description="SO4 concentration"

reminOrgOxO2 (mol O2 yr-1), VT_ReactProperty, description="oxygen consumption (-ve) by remineralization"

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactProperty, description="2 * sulphate consumption (-ve) by remineralization"

reminOrgOxCH4 (mol O2eq yr-1), VT_ReactProperty, description="2 * DIC -> methane (-ve) by remineralization"

RateStoich_reminOrgOxO2

reminOrgOxO2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

[soluteflux_O2] (), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxO2"

RateStoich_reminOrgOxSO4

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactDependency, description="2 * sulphate consumption (-ve) by remineralization"

[soluteflux_SO4] (), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

[soluteflux_H2S] (), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

[soluteflux_TAlk] (), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxSO4"

RateStoich_reminOrgOxCH4

reminOrgOxCH4 (mol O2eq yr-1), VT_ReactDependency, description="2 * DIC -> methane (-ve) by remineralization"

[soluteflux_DIC] (), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxCH4"

[soluteflux_CH4] (), VT_ReactContributor, description="generated by RateStoich rate=reminOrgOxCH4"

totals

reminOrgOxSO4 (mol O2eq yr-1), VT_ReactDependency, description="2 * sulphate consumption (-ve) by remineralization"

reminOrgOxO2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (-ve) by remineralization"

reminOrgOxCH4 (mol O2eq yr-1), VT_ReactDependency, description="2 * DIC -> methane (-ve) by remineralization"

reminOrgOxSO4_total (mol O2eq yr-1), VT_ReactProperty, description="total 2 * sulphate consumption (-ve) by remineralization"

reminOrgOxO2_total (mol O2 yr-1), VT_ReactProperty, description="total oxygen consumption (-ve) by remineralization"

reminOrgOxCH4_total (mol O2eq yr-1), VT_ReactProperty, description="total 2 * DIC -> methane (-ve) by remineralization"

Secondary redox

PALEOaqchem.SecondaryRedox.ReactionRedoxH2S_O2 — Type

ReactionRedoxH2S_O2

Sulphide oxidation by oxygen.

Rate R_H2S_O2	Units	Ref	Notes

1.6e5	(mol l-1)-1 yr-1	Philippe {Van Cappellen}, Y. Wang (1996)

3.65e6	(mol l-1)-1 yr-1	Kazumi Ozaki, Shigeo Tajima, Eiichi Tajika (2011)

54e6	(mol l-1)-1 yr-1	Stephen J. Romaniello, Louis A Derry (2010)

Parameters

R_H2S_O2[Float64]=3650.0 ((mol m-3)-1 yr-1), default_value=3650.0, description="rate constant"

SIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable sulphur isotopes and specify isotope type"

Methods and Variables for default Parameters

do_redox_H2S_O2

redox_H2S_O2 (mol O2 yr-1), VT_ReactProperty, description="oxygen consumption (+ve) by H2S oxidation"

O2_conc (mol m-3), VT_ReactDependency, description="O2 concentration"

H2S_conc (mol m-3), VT_ReactDependency, description="H2S concentration"

volume (m3), VT_ReactDependency, description="box fluid volume"

RateStoich_redox_H2S_O2

redox_H2S_O2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (+ve) by H2S oxidation"

[O2_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_H2S_O2"

[H2S_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_H2S_O2"

[SO4_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_H2S_O2"

[TAlk_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_H2S_O2"

totals

redox_H2S_O2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (+ve) by H2S oxidation"

redox_H2S_O2_total (mol O2 yr-1), VT_ReactProperty, description="total oxygen consumption (+ve) by H2S oxidation"

PALEOaqchem.SecondaryRedox.ReactionRedoxCH4_O2 — Type

ReactionRedoxCH4_O2

Methane oxidation by oxygen.

Rate R_CH4_O2	Units	Ref	Notes

1e10	(mol l-1)-1 yr-1	Philippe {Van Cappellen}, Y. Wang (1996)

10e6	(mol l-1)-1 yr-1	TODO

Parameters

R_CH4_O2[Float64]=10000.0 ((mol m-3)-1 yr-1), default_value=10000.0, description="rate constant"

CIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable carbon isotopes and specify isotope type"

Methods and Variables for default Parameters

do_redox_CH4_O2

redox_CH4_O2 (mol O2 yr-1), VT_ReactProperty, description="oxygen consumption (+ve) by CH4 oxidation"

O2_conc (mol m-3), VT_ReactDependency, description="O2 concentration"

CH4_conc (mol m-3), VT_ReactDependency, description="CH4 concentration"

volume (m3), VT_ReactDependency, description="box fluid volume"

RateStoich_redox_CH4_O2

redox_CH4_O2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (+ve) by CH4 oxidation"

[O2_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_CH4_O2"

[CH4_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_CH4_O2"

[DIC_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_CH4_O2"

totals

redox_CH4_O2 (mol O2 yr-1), VT_ReactDependency, description="oxygen consumption (+ve) by CH4 oxidation"

redox_CH4_O2_total (mol O2 yr-1), VT_ReactProperty, description="total oxygen consumption (+ve) by CH4 oxidation"

PALEOaqchem.SecondaryRedox.ReactionRedoxCH4_SO4 — Type

ReactionRedoxCH4_SO4

Methane oxidation by sulphate (anaerobic methane oxidation).

Rate R_CH4_SO4	Units	Ref	Notes

1e4	(mol l-1)-1 yr-1	Philippe {Van Cappellen}, Y. Wang (1996)	k17

Parameters

R_CH4_SO4[Float64]=10.0 ((mol m-3)-1 yr-1), default_value=10.0, description="rate constant"

CIsotope[external, DataType]=PALEOboxes.ScalarData, default_value=PALEOboxes.ScalarData, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable carbon isotopes and specify isotope type"

Methods and Variables for default Parameters

do_redox_CH4_SO4

redox_CH4_SO4 (mol CH4/SO4 yr-1), VT_ReactProperty, description="sulphate consumption (+ve) by CH4 oxidation"

SO4_conc (mol m-3), VT_ReactDependency, description="SO4 concentration"

CH4_conc (mol m-3), VT_ReactDependency, description="CH4 concentration"

volume (m3), VT_ReactDependency, description="box fluid volume"

RateStoich_redox_CH4_SO4

redox_CH4_SO4 (mol CH4/SO4 yr-1), VT_ReactDependency, description="sulphate consumption (+ve) by CH4 oxidation"

[SO4_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_CH4_SO4"

[CH4_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_CH4_SO4"

[DIC_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_CH4_SO4"

[H2S_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_CH4_SO4"

[TAlk_sms] (), VT_ReactContributor, description="generated by RateStoich rate=redox_CH4_SO4"

totals

redox_CH4_SO4 (mol CH4/SO4 yr-1), VT_ReactDependency, description="sulphate consumption (+ve) by CH4 oxidation"

redox_CH4_SO4_total (mol CH4/SO4 yr-1), VT_ReactProperty, description="total sulphate consumption (+ve) by CH4 oxidation"

Carbonate chemistry

PALEOaqchem.CarbChem.ReactionCO2SYS — Type

ReactionCO2SYS

Carbonate chemistry using paleo_CO2SYS.

Three different pH solver approaches are supported, set by Parameter solve_pH:

solve: iterative solution for pH, TAlk is a StateExplicit Variable (eg using explicit ODE solver)

constraint: additional State Variable for pH and algebraic Constraint for TAlk (requires DAE solver).

implicit: TAlk is a Total Variable, a function of a State Variable for pH (requires DAE solver)

Parameters

WhichKs[Int64]=10, default_value=10, description="PALEOco2sys WhichKs choice of eqb constant data"

WhoseKSO4[Int64]=1, default_value=1, description="PALEOco2sys WhoseKSO4 choice of eqb constant data"

components[Vector{String}]=["Ci", "S", "F", "B", "Omega"], default_value=["Ci", "S", "F", "B", "Omega"], allowed_values=["Ci", "S", "F", "P", "B", "Si", "H2S", "NH3", "Omega"], description="PALEOco2sys optional contributions to TAlk"

defaultconcs[Vector{String}]=["TS", "TF", "TB", "Ca"], default_value=["TS", "TF", "TB", "Ca"], allowed_values=["TS", "TF", "TB", "Ca"], description="PALEOco2sys default concentrations (modern values calculated from salinity)"

output_consts[Vector{String}]=String[], default_value=String[], description="PALEOco2sys eqb constants etc to include as output variables"

outputs[Vector{String}]=["pCO2", "xCO2dryinp"], default_value=["pCO2", "xCO2dryinp"], description="PALEOco2sys output concentrations etc to include as output variables"

output_pHtot[Bool]=true, default_value=true, description="true to output pHtot (requires TS and TF)"

solve_pH[String]="solve", default_value="solve", allowed_values=["solve", "constraint", "implicit"], description="'solve' to solve iteratively for pH, 'constraint' to include [pHfree] as a state variable and TAlk as a constraint, 'implicit' to include [pHfree] as a state variable and TAlk an implicit variable"

pHtol[Float64]=2.220446049250313e-14, default_value=2.220446049250313e-14, description="pH tolerance for iterative solution"

simd_width[String]="1", default_value="1", allowed_values=["1", "FP64P2", "FP64P4", "FP32P4", "FP32P8"], description="use SIMD ("1" - no SIMD, "FP64P4" - 4 x Float64, etc)"

Methods and Variables for default Parameters

do_carbchem_simd

volume (m^3), VT_ReactDependency, description="Cell volume"

rho_ref (kg m^-3), VT_ReactDependency, description="density conversion factor"

temp (Kelvin), VT_ReactDependency, description="temperature"

pressure (dbar), VT_ReactDependency, description="pressure"

sal (psu), VT_ReactDependency, description="salinity"

pHfree (), VT_ReactProperty, description="-log10(hydrogen ion concentration)"

TAlk_conc (mol m^-3), VT_ReactDependency, description="total TAlk concentration"

pHtot (), VT_ReactProperty, description="-log10([H] + [HS]), pH on total scale"

TS_conc –> %reaction%TS_conc (mol m^-3), VT_ReactProperty, description="TS modern default scaled from salinity"

TF_conc –> %reaction%TF_conc (mol m^-3), VT_ReactProperty, description="TF modern default scaled from salinity"

TB_conc –> %reaction%TB_conc (mol m^-3), VT_ReactProperty, description="TB modern default scaled from salinity"

Ca_conc –> %reaction%Ca_conc (mol m^-3), VT_ReactProperty, description="Ca modern default scaled from salinity"

TCi_conc (mol m^-3), VT_ReactDependency, description="TCi concentration"

pCO2 –> %reaction%pCO2 (atm), VT_ReactProperty, description="CO2 partial pressure"

xCO2dryinp –> %reaction%xCO2dryinp (), VT_ReactProperty, description="mixing ratio of CO2 in dry air at 1 atm (always > pCO2 due to H2O vapour pressure)"

Isotope systems

Boron

PALEOaqchem.Boron.ReactionBoronIsotope — Type

ReactionBoronIsotope

Calculate d11B for aqueous B(OH)4- and B(OH)3 species from mass balance, given d11B for total B, total B concentration B_conc and B(OH)4- concentration BOH4_conc.

See eg Richard E Zeebe, Dieter a. Wolf-Gladrow (2001), p220.

Parameters

alphaB[Float64]=1.0272, default_value=1.0272, description="isotopic fractionation factor B(OH)4m <-> B(OH)3"

BIsotope[external, UnionAll]=PALEOboxes.IsotopeLinear, default_value=PALEOboxes.IsotopeLinear, allowed_values=Type[PALEOboxes.ScalarData, PALEOboxes.IsotopeLinear], description="disable / enable boron isotopes and specify isotope type"

Methods and Variables

do_boron_isotope

B_conc (mol m-3), VT_ReactDependency, description="B concentration"

BOH4_conc (mol m-3), VT_ReactDependency, description="B(OH)4m concentration"

B_delta (per mil), VT_ReactDependency, description="d11B delta for total B"

BOH4_delta (per mil), VT_ReactProperty, description="d11B delta for B(OH)4- species"